requirements.txt
getmac==0.8.2
paho-mqtt==1.5.1
